Fixed (ActivityController): Changed type int to long for activityId (#244)

+++ GirafRest/Controllers/ActivityController.cs
@@ -197,7 +197,7 @@ namespace GirafRest.Controllers
         [HttpGet("{userId}/{activityId}")]
         [Authorize]
         [ProducesResponseType(StatusCodes.Status200OK)]
-        public async Task<ActionResult> GetActivity(string userId, int activityId)
+        public async Task<ActionResult> GetActivity(string userId, long activityId)
         {
             var activity = _activityRepository.Get(activityId);
             var pictograms = _pictogramRelationRepository.GetWithPictogram(activityId);
